+++ packages/field/styles.css
@@ -1,4 +1,5 @@
 @import "@pluralsight/ps-design-system-icon/styles.css";
+@import "@pluralsight/ps-design-system-halo/styles.css";
 @import "./dist/esm/css/field.css";
 @import "./dist/esm/css/input.css";
 @import "./dist/esm/css/label.css";
