+++ packages/menu/package.json
@@ -37,6 +37,7 @@
   "types": "dist/esm/index.d.ts",
   "dependencies": {
     "@pluralsight/ps-design-system-core": "^9.0.1",
+    "@pluralsight/ps-design-system-icon": "^24.1.1",
     "@pluralsight/ps-design-system-util": "^9.3.1",
     "focus-within": "^3.0.2"
   },
@@ -47,7 +48,6 @@
   },
   "devDependencies": {
     "@pluralsight/ps-design-system-build": "^4.0.0",
-    "@pluralsight/ps-design-system-icon": "^24.1.1",
     "@pluralsight/ps-design-system-position": "^8.2.3",
     "@pluralsight/ps-design-system-storybook-preset": "^0.7.6"
   },
