+++ packages/menu/src/index.ts
@@ -1,3 +1,4 @@
+import '@pluralsight/ps-design-system-icon/styles.css'
 import './css/index.css'
 import Menu from './react/index'
 export default Menu
