+++ packages/select/package.json
@@ -37,6 +37,7 @@
   "types": "dist/esm/index.d.ts",
   "dependencies": {
     "@pluralsight/ps-design-system-core": "^9.0.1",
+    "@pluralsight/ps-design-system-icon": "^24.1.1",
     "@pluralsight/ps-design-system-menu": "^4.0.3",
     "@pluralsight/ps-design-system-util": "^9.3.1"
   },
