+++ packages/select/src/index.ts
@@ -1,4 +1,5 @@
-import '@pluralsight/ps-design-system-menu/style.css'
+import '@pluralsight/ps-design-system-icon/styles.css'
+// import '@pluralsight/ps-design-system-menu/styles.css'
 import './css/index.css'
 import Select from './react/index'
 export default Select
