+++ packages/switch/src/index.ts
@@ -1,3 +1,4 @@
+import '@pluralsight/ps-design-system-halo/styles.css'
 import './css/index.css'
 import Switch from './react/index'
 export default Switch
