+++ packages/tag/src/index.ts
@@ -1,4 +1,5 @@
 import '@pluralsight/ps-design-system-icon/styles.css'
+import '@pluralsight/ps-design-system-halo/styles.css'
 import './css/index.css'
 import Tag from './react/index'
 export default Tag
